[mlir][scf] Add a pass to expand scf.if regions
This pass expands the scf.if op regions by pulling in ops before
and after the scf.if op into both regions. This is an useful
transformation to put ops within the same region so that we
can optimze together better, e.g., when we have a fast-slow path
in the same kernel we'd want to optimize each path separately.